requirements.txt
amqp==5.0.9
APScheduler==3.6.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.7.0
billiard==3.6.4.0
build==0.10.0
cachetools==4.2.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.1.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==3.6
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.8
django-celery-beat==2.2.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.2.0
django-filter==22.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
djangorestframework-xml==2.0.0
docker==5.0.3
fpdf==1.7.2
idna==3.3
iniconfig==1.1.1
kombu==5.2.3
numpy==1.24.2
packaging==21.3
pandas==2.0.0
Pillow==9.2.0
pluggy==1.0.0
prompt-toolkit==3.0.27
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.9.1
PyJWT==2.3.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
reportlab==3.6.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tornado==6.1
tzdata==2023.3
tzlocal==4.1
unicode==2.9
Unidecode==1.3.6
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.3
whitenoise==5.3.0
wrapt==1.14.1
